requirements.txt
elsa>=0.1.2
markdown
pyyaml
ansicolor
